delete images and add album work properly now

+++ worldpix/src/app/components/image-cascade/image-cascade.component.ts
@@ -1,28 +1,34 @@
-import { Component, Input } from '@angular/core';
+import { Component, EventEmitter, Input, Output } from '@angular/core';
 import { ImageService } from '../../services/image.service';
 import { Image } from '../../interfaces/image';
 import { NgFor } from '@angular/common';
 import { IconRoundButtonComponent } from "../buttons/icon-round-button/icon-round-button.component";
+import { ConfirmPopupComponent } from "../confirm-popup/confirm-popup.component";
 
 @Component({
     selector: 'app-image-cascade',
     standalone: true,
     templateUrl: './image-cascade.component.html',
     styleUrl: './image-cascade.component.css',
-    imports: [NgFor, IconRoundButtonComponent]
+    imports: [NgFor, IconRoundButtonComponent, ConfirmPopupComponent]
 })
 export class ImageCascadeComponent {
   constructor(private imageService: ImageService) {}
   
   images: Image[] = [];
   @Input() albumId: string = '';
+  @Output() imageDeleted: EventEmitter<any> = new EventEmitter();
+
 
   ngAfterViewInit() {
     this.loadImages();
   }
 
-  deleteImage(image: Image) {
-    this.imageService.deleteImage(image.id!);
+  deleteImage(result: boolean, image: Image) {
+    if (result) {
+      this.imageService.deleteImage(image.id!);
+      this.imageDeleted.emit()
+    }
   }
 
   loadImages() {
